# Conflicts:
#	restoMobile/src/pages/ProfileScreen/Login/Login.tsx
#	visitorMobile/src/pages/ProfileScreen/Login/Login.tsx

+++ restoMobile/src/services/restoCalls.ts
@@ -3,9 +3,13 @@ import axios from 'axios';
 import { API_URL } from '@env';
 
 const baseURL = API_URL;
+const menuDesignUrl = `${API_URL}menuDesigns/`;
 
 export const getAllResto = async () => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: 'GET',
       url: baseURL + 'restaurants/',
@@ -19,6 +23,9 @@ export const getAllResto = async () => {
 
 export const getAllRestaurantsByUser = async (body: any) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: "GET",
       url: baseURL + 'restaurants/user/resto',
@@ -34,8 +41,31 @@ export const getAllRestaurantsByUser = async (body: any) => {
   }
 };
 
+export const getAllRestaurantsByUserAndFilter = async (token: string, filter: string) => {
+  try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
+    const response = await axios({
+      method: "POST",
+      url: baseURL + 'search/restaurants/',
+      data: { "token": token,"filter": filter },
+      headers: {
+        "content-type": "application/json",
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error("Error fetching all restaurants:", error);
+    throw new Error("Failed to fetch all restaurants");
+  }
+};
+
 export const getRestaurantByName = async(name: string) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: 'GET',
       url: baseURL + 'restaurants/'+name,
@@ -51,11 +81,30 @@ export const getRestaurantByName = async(name: string) => {
 }
 
 export const addRestaurant = async(restaurantData: any) => {
-  return (await axios.post('${API_URL}restaurants/', restaurantData));
+  try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
+    const response = await axios({
+      method: 'POST',
+      url: baseURL + 'restaurants/',
+      data: restaurantData,
+      headers: {
+        'content-type': 'application/json',
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error('Error adding restaurant:', error);
+    throw new Error('Failed to add restaurant');
+  }
 }
 
 export const deleteRestaurantByName = async (restaurantName: string) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     await axios({
       method: 'DELETE',
       url: baseURL +'restaurants/' + restaurantName,
@@ -63,5 +112,86 @@ export const deleteRestaurantByName = async (restaurantName: string) => {
   } catch (error) {
     console.error('Error deleting restaurant:', error);
     throw new Error('Failed to delete restaurant');
+  }
+};
+
+export const getAllMenuDesigns = async () => {
+  try {
+    if (menuDesignUrl === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
+    const response = await axios({
+      method: "GET",
+      url: menuDesignUrl
+    });
+    return response.data;
+  } catch (error) {
+    console.error("Error fetching all menu designs:", error);
+    throw new Error("Failed to fetch all menu designs");
+  }
+};
+
+export const editResto = async (restoName: string, body: any) => {
+  try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
+    const response = await axios({
+      url: baseURL + 'restaurants/' + restoName,
+      method: "PUT",
+      data: JSON.stringify(body),
+      headers: {
+        "content-type": "application/json",
+      },
+    });
+
+    if (response.status === 200) {
+      return response.data;
+    } else {
+      return false;
+    }
+  } catch (error) {
+    console.error("Error editing restaurant:", error);
+    throw new Error("Failed to edit restaurant");
+  }
+};
+
+export const getRestoByName = async (restoName: string) => {
+  try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
+    const response = await axios({
+      url: baseURL + 'restaurants/' + restoName,
+      method: "GET",
+      headers: {
+        "content-type": "application/json",
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error("Error editing restaurant:", error);
+    throw new Error("Failed to edit restaurant");
+  }
+};
+
+export const updateRestoCategories = async (userToken: string, uid: number, newCategories: any) => {
+  try {
+    const response = await axios({
+      url: baseURL + 'restaurants/updateCategories', 
+      method: 'POST',
+      data: {
+        userToken: userToken,
+        uid: uid,
+        newCategories: newCategories
+      },
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+    return response.data;
+  } catch (error) {
+    console.error('Error updating user categories:', error);
+    throw new Error('Failed to update user categories');
   }
 };
